Move progress_range to helpers, add progress parameter to emcee interface
