+++ nuts/helpers.py
@@ -6,7 +6,10 @@ _function_wrapper           hack to make partial functions pickleable
 NutsSampler_fn_wrapper      combine provided lnp and grad(lnp) into one function
 """
 import numpy as np
-
+try:
+    import tqdm
+except ImportError:
+    tqdm = None
 
 def numerical_grad(theta, f, dx=1e-3, order=1):
     """ return numerical estimate of the local gradient
@@ -136,3 +139,29 @@ class NutsSampler_fn_wrapper(object):
         if self.verbose:
             print(r[0], theta)
         return r
+
+def progress_range(minimum, maximum, progress=True):
+    """A range-like function which displays progress information.
+
+    INPUTS
+    ------
+    minimum: int
+        Lower bound of range
+    maximum: int
+        Upper bound of range
+    KEYWORDS
+    --------
+    progress: bool
+        If True, show progress, otherwise display nothing
+    """
+    if not progress:
+        for i in range(minimum, maximum):
+            yield i
+    elif tqdm is not None:
+        for i in tqdm.trange(minimum, maximum):
+            yield i
+    else:
+        for i in range(minimum, maximum):
+            if i % 100 == 0:
+                print('iteration %i/%i' % (i, maximum-minimum))
+            yield i
